pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["kaustubh <kaustubhkagrawal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = {extras = ["async"], version = "^3.0.2"}
openai = "^1.11.1"
flask-cors = "^4.0.0"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
llama-index = "^0.9.19"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"